Transferring over new updates - Scott 7/2/2017

+++ Test_code/error_checks/input_file_parser.py
@@ -108,7 +108,7 @@ def generateTemplate(source, filename="input_file", sumname="summary_file", data
 
 		####obtain number of particles
 		particles = particles_regex.search(info).group(1)
-		particles = float(particles)
+		particles = int(particles)
 		####
